ngoding-seru/Spoj-Solutions-Downloader-master/requirements.txt
beautifulsoup4==4.5.1
html5==0.0.9
requests==2.11.1
robobrowser==0.5.3
six==1.10.0
webencodings==0.5
Werkzeug==0.11.11
